+++ bpod_core/bpod.py
@@ -2,317 +2,245 @@ from __future__ import annotations
 
 import logging
 from abc import ABC, abstractmethod
-from collections.abc import Iterator
-from typing import TYPE_CHECKING, Any, NamedTuple
+from typing import TYPE_CHECKING, NamedTuple
 
-import serial
+from pydantic import validate_call
 from serial import SerialException
-from serial.threaded import Protocol, ReaderThread
-from serial.tools import list_ports
+from serial.tools.list_ports import comports
 
-from bpod_core import __version__ as VERSION  # noqa: N812
-from bpod_core.serial_extensions import (
-    SerialSingleton,
-    SerialSingletonException,
-    get_serial_number_from_port,
-)
+from bpod_core import __version__ as bpod_core_version
+from bpod_core.serial import ExtendedSerial
 
 if TYPE_CHECKING:
     from _typeshed import ReadableBuffer  # noqa: F401
 
 PROJECT_NAME = 'bpod-core'
-VID_TEENSY = 0x16C0
+VIDS_BPOD = [0x16C0]  # vendor IDs of supported Bpod devices
+MIN_BPOD_FW_VERSION = (23, 0)  # minimum supported firmware version (major, minor)
+MIN_BPOD_HW_VERSION = 3  # minimum supported hardware version
+MAX_BPOD_HW_VERSION = 4  # maximum supported hardware version
 
 logger = logging.getLogger(__name__)
 
 
-class SerialReaderProtocolRaw(Protocol):
-    def connection_made(self, transport):
-        """
-        Called when the reader thread is started.
+class VersionInfo(NamedTuple):
+    """Represents the Bpod's on-board hardware configuration."""
 
-        Parameters
-        ----------
-        - transport: The transport object associated with the connection.
-        """
-        logger.info('Threaded serial reader started - ready to receive data ...')
-
-    def data_received(self, data):
-        """
-        Called with snippets received from the serial port.
-
-        Parameters
-        ----------
-        - data: The binary data received from the serial port.
-        """
-        print(data.decode())
-
-    def connection_lost(self, exc):
-        """
-        Called when the connection is lost or encounters an exception.
-
-        Parameters
-        ----------
-        - exc: The exception that caused the connection loss, if any.
-        """
-        logger.info(exc)
+    firmware: tuple[int, int]
+    """Firmware version (major, minor)"""
+    machine: int
+    """Machine type (numerical)"""
+    pcb: int | None
+    """PCB revision, if applicable"""
 
 
-class BpodException(SerialSingletonException):
-    pass
+class HardwareConfiguration(NamedTuple):
+    """Represents the Bpod's on-board hardware configuration."""
+
+    max_states: int
+    """Maximum number of supported states in a single state machine description."""
+    timer_period: int
+    """Period of the state machine's refresh cycle during a trial in microseconds."""
+    max_serial_events: int
+    """Maximum number of behavior events allocatable among connected modules."""
+    max_bytes_per_serial_message: int
+    """Maximum number of bytes allowed per serial message."""
+    n_global_timers: int
+    """Number of global timers supported."""
+    n_global_counters: int
+    """Number of global counters supported."""
+    n_conditions: int
+    """Number of condition-events supported."""
+    n_inputs: int
+    """Number of input channels."""
+    input_description: bytes
+    """Array indicating the state machine's onboard input channel types."""
+    n_outputs: int
+    """Number of channels in the state machine's output channel description array."""
+    output_description: bytes
+    """Array indicating the state machine's onboard output channel types."""
 
 
-class Bpod(SerialSingleton):
+class BpodError(Exception):
     """
-    Class for interfacing a Bpod Finite State Machine.
+    Exception class for Bpod-related errors.
 
-    The Bpod class extends :class:`serial.Serial`.
-
-    Parameters
-    ----------
-    port : str, optional
-        The serial port for the Bpod device, or None to automatically detect a Bpod.
-    connect : bool, default: True
-        Whether to connect to the Bpod device. If True and 'port' is None, an
-        attempt will be made to automatically find and connect to a Bpod device.
-    **kwargs
-        Additional keyword arguments passed to :class:`serial.Serial`.
-
-    Examples
-    --------
-    * Try to automatically find a Bpod device and connect to it.
-
-        .. code-block:: python
-
-            my_bpod = Bpod()
-
-    * Connect to a Bpod device on COM3
-
-        .. code-block:: python
-
-            my_bpod = Bpod('COM3')
-
-    * Instantiate a Bpod object for a device on COM3 but only connect to it later.
-
-        .. code-block:: python
-
-            my_bpod = Bpod(port = "COM3", connect = False)
-            # (do other things)
-            my_bpod.open()
+    This exception is raised when an error specific to the Bpod device or its
+    operations occurs.
     """
 
-    class _Info(NamedTuple):
-        serial_number: str
-        firmware_version: tuple[int, int]
-        machine_type: int
-        machine_type_string: str
-        pcb_revision: int
-        max_states: int
-        timer_period: int
-        max_serial_events: int
-        max_bytes_per_serial_message: int
-        n_global_timers: int
-        n_global_counters: int
-        n_conditions: int
-        n_inputs: int
-        input_description_array: bytes
-        n_outputs: int
-        output_description_array: bytes
 
-    def __new__(
-        cls,
-        port: str | None = None,
-        connect: bool = True,
-        **kwargs,
-    ):
+class Bpod:
+    """Bpod class for interfacing with the Bpod Finite State Machine."""
+
+    _version: VersionInfo
+    _hardware_config: HardwareConfiguration
+    serial0: ExtendedSerial
+    """Primary serial device for communication with the Bpod."""
+    serial1: ExtendedSerial
+    """Secondary serial device for communication with the Bpod."""
+    serial2: ExtendedSerial | None = None
+    """Tertiary serial device for communication with the Bpod - used by Bpod 2+ only."""
+
+    @validate_call
+    def __init__(self, port: str | None = None, serial_number: str | None = None):
+        logger.info(f'bpod_core {bpod_core_version}')
+
+        # identify Bpod by port or serial number
+        port, serial_number = self._identify_bpod(port, serial_number)
+
+        # open primary serial port
+        self.serial0 = ExtendedSerial()
+        self.serial0.port = port
+        self.open()
+
+        # get firmware version and machine type; enforce version requirements
+        self._get_version_info()
+
+        # get the Bpod's onboard hardware configuration
+        self._get_hardware_configuration()
+
+        # configure input and output channels
+        self._configure_channels()
+
+        # detect additional serial ports
+        self._detect_additional_serial_ports()
+
+        # log hardware information
+        machine = {3: 'r2.0-2.5', 4: '2+ r1.0'}.get(self.version.machine, 'unknown')
+        logger.info(f'Connected to Bpod Finite State Machine {machine} on {self.port}')
+        logger.info(
+            f'Firmware Version {"{}.{}".format(*self.version.firmware)}, '
+            f'Serial Number {serial_number}, PCB Revision {self.version.pcb}'
+        )
+
+    def __enter__(self):
+        """Enter context."""
+        return self
+
+    def __exit__(self, type, value, traceback):
+        """Exit context and close connection."""
+        self.close()
+
+    def __del__(self):
+        self.close()
+
+    @staticmethod
+    def _identify_bpod(
+        port: str | None = None, serial_number: str | None = None
+    ) -> tuple[str, str | None]:
         """
-        Create or retrieve a singleton instance of the Bpod class.
+        Try to identify a supported Bpod based on port or serial number.
 
-        This method implements a singleton pattern for the Bpod class, ensuring that
-        only one instance is created for a given port. If an instance already exists
-        for the specified port, that instance is returned.
+        If neither port nor serial number are provided, this function will attempt to
+        detect a supported Bpod automatically.
 
         Parameters
         ----------
-        port : str, optional
-            The serial port for the Bpod device, or None to automatically detect a Bpod.
-        connect : bool, optional
-            Whether to connect to the Bpod device. If True and 'port' is None, an
-            attempt will be made to automatically find and connect to a Bpod device.
-        **kwargs
-            Additional keyword arguments passed to serial.Serial.
+        port : str | None, optional
+            The port of the device.
+        serial_number : str | None, optional
+            The serial number of the device.
 
         Returns
         -------
-        Bpod
-            A singleton instance of the Bpod class.
+        str
+            the port of the device
+        str | None
+            the serial number of the device
 
         Raises
         ------
-        ValueError
-            If 'port' is not a string and is not None.
-
-        Notes
-        -----
-        The singleton instances are managed by a class-level lock and dictionary.
-        Automatic Bpod detection relies on the find method.
-
-        Example
-        -------
-        To create or retrieve a Bpod instance on a specific port:
-
-        .. code-block:: python
-            bpod_instance = Bpod(port='COM3')
-
-        To automatically detect and create or retrieve a Bpod instance:
-
-        .. code-block:: python
-            bpod_instance = Bpod()
+        BpodError
+            If no Bpod is found or the indicated device is not supported.
         """
-        # log version
-        logger.debug(f'{PROJECT_NAME} {VERSION}')
 
-        # try to automagically find a Bpod device
-        if port is None and connect is True:
-            port = next(iter(Bpod._instances.keys()), next(find_bpod_ports(), None))
+        def sends_discovery_byte(port: str) -> bool:
+            """Check if the device on the given port sends a discovery byte."""
+            try:
+                with ExtendedSerial(port, timeout=0.15) as ser:
+                    return ser.read(1) == bytes([222])
+            except SerialException:
+                return False
 
-        # implement singleton
-        return super().__new__(cls, port, **kwargs)
+        # If no port or serial number provided, try to automagically find an idle Bpod
+        if port is None and serial_number is None:
+            try:
+                port_info = next(
+                    p
+                    for p in comports()
+                    if getattr(p, 'vid', None) in VIDS_BPOD
+                    and sends_discovery_byte(p.device)
+                )
+            except StopIteration as e:
+                raise BpodError('No available Bpod found') from e
+            return port_info.device, port_info.serial_number
 
-    def __init__(self, port: str | None = None, connect: bool = True, **kwargs) -> None:
+        # Else, if a serial number was provided, try to match it with a serial device
+        elif serial_number is not None:
+            try:
+                port_info = next(
+                    p
+                    for p in comports()
+                    if p.serial_number == serial_number
+                    and sends_discovery_byte(p.device)
+                )
+            except (StopIteration, AttributeError) as e:
+                raise BpodError(f'No device with serial number {serial_number}') from e
+
+        # Else, assure that the provided port exists and the device could be a Bpod
+        else:
+            try:
+                port_info = next(p for p in comports() if p.device == port)
+            except (StopIteration, AttributeError) as e:
+                raise BpodError(f'Port not found: {port}') from e
+
+        if port_info.vid not in VIDS_BPOD:
+            raise BpodError('Device is not a supported Bpod')
+        return port_info.device, port_info.serial_number
+
+    def _get_version_info(self) -> None:
         """
-        Initialize a Bpod instance.
+        Retrieve firmware version and machine type information from the Bpod.
 
-        This method initializes a Bpod instance, allowing communication with a Bpod
-        device over a specified serial port.
-
-        Parameters
-        ----------
-        port : str, optional
-            The serial port for the Bpod device. If None and 'connect' is True, an
-            attempt will be made to automatically detect and use a Bpod port.
-        connect : bool, optional
-            Whether to establish a connection to the Bpod device. If True and 'port' is
-            None, automatic port detection will be attempted.
-        **kwargs
-            Additional keyword arguments to be passed to the constructor of
-            serial.Serial.
-
-        Notes
-        -----
-        -   If the Bpod instance is already instantiated, the method returns without
-            further action.
-        -   If 'port' is 'None' and 'connect' is True the former value may be
-            overridden based on existing instances
-        """
-        if self._initialized:
-            return
-
-        self.info: Bpod._Info | None = None
-        self.inputs = None
-        self.outputs = None
-        self._reader: ReaderThread = ReaderThread(self, SerialReaderProtocolRaw)
-
-        # automatic port discovery (also see __new__)
-        if port is None and connect is True:
-            port = next((k for (k, v) in self._instances.items() if v is self), None)
-
-        # initialize super class
-        if 'baudrate' not in kwargs:
-            kwargs['baudrate'] = 1312500
-        super().__init__(port=port, connect=connect, **kwargs)
-        assert self._initialized is True
-
-    def __repr__(self):
-        return f'Bpod(port={self.port})'
-
-    def open(self) -> None:
-        """
-        Open serial connection and connect to Bpod Finite State Machine.
+        This method queries the Bpod to obtain its firmware version, machine type, and
+        PCB revision. It also validates that the hardware and firmware versions meet
+        the minimum requirements. If the versions are not supported, an Exception is
+        raised.
 
         Raises
         ------
-        BpodException
-            Handshake failed: The Bpod did not acknowledge our request.
+        BpodError
+            If the hardware version or firmware version is not supported.
         """
-        super().open()
-
-        # try to perform handshake
-        if self.handshake():
-            logger.debug('Handshake successful')
-
-        # get firmware version and machine type; assert version requirements
-        v_major, machine_type = self.query(b'F', '<2H')
-        version = (v_major, self.query(b'f', '<H')[0] if v_major > 22 else 0)
-        if not (2 < machine_type < 5):
-            raise BpodException(
+        v_major, machine_type = self.serial0.query_struct(b'F', '<2H')
+        v_minor = self.serial0.query_struct(b'f', '<H')[0] if v_major > 22 else 0
+        v_firmware = (v_major, v_minor)
+        if not (MIN_BPOD_HW_VERSION <= machine_type <= MAX_BPOD_HW_VERSION):
+            raise BpodError(
                 f'The hardware version of the Bpod on {self.port} is not supported.'
             )
-        if version < (min_version := (23, 0)):
-            raise BpodException(
-                f'The Bpod on {self.port} uses firmware v{version[0]}.{version[1]} '
+        if v_firmware < MIN_BPOD_FW_VERSION:
+            raise BpodError(
+                f'The Bpod on {self.port} uses firmware v{v_major}.{v_minor} '
                 f'which is not supported. Please update the device to '
-                f'firmware v{min_version[0]}.{min_version[1]} or later.'
+                f'firmware v{MIN_BPOD_FW_VERSION[0]}.{MIN_BPOD_FW_VERSION[1]} or later.'
             )
+        pcv_rev = self.serial0.query_struct(b'v', '<B')[0] if v_major > 22 else None
+        self._version = VersionInfo(v_firmware, machine_type, pcv_rev)
 
-        # detect additional USB-serial ports
-        candidate_ports = [
-            p.device
-            for p in list_ports.comports()
-            if p.vid == VID_TEENSY and p.device != self.port
-        ]
-        for port in candidate_ports:
-            try:
-                with serial.Serial(port, timeout=0.2) as ser:
-                    if ser.read(1) == bytes([222]):
-                        candidate_ports.remove(port)
-            except serial.SerialException:
-                pass
-        for port in candidate_ports:
-            try:
-                with serial.Serial(port, timeout=0.2) as ser:
-                    self.write(b'{')
-                    if ser.read(1) == bytes([222]):
-                        print(port)
-                        candidate_ports.remove(port)
-                        continue
-            except serial.SerialException:
-                pass
-        if machine_type == 4:
-            for port in candidate_ports:
-                try:
-                    with serial.Serial(port, timeout=0.2) as ser:
-                        self.write(b'}')
-                        if ser.read(1) == bytes([223]):
-                            print(port)
-                            continue
-                except serial.SerialException:
-                    pass
+    def _get_hardware_configuration(self) -> None:
+        """Retrieve the Bpod's onboard hardware configuration."""
+        if self.version.firmware > (22, 0):
+            hardware_conf = list(self.serial0.query_struct(b'H', '<2H6B'))
+        else:
+            hardware_conf = list(self.serial0.query_struct(b'H', '<2H5B'))
+            hardware_conf.insert(-4, 3)  # max bytes per serial msg always = 3
+        hardware_conf.extend(self.serial0.read_struct(f'<{hardware_conf[-1]}s1B'))
+        hardware_conf.extend(self.serial0.read_struct(f'<{hardware_conf[-1]}s'))
+        self._hardware_config = HardwareConfiguration(*hardware_conf)
 
-        # get some more hardware information
-        machine_str = {3: 'r2.0-2.5', 4: '2+ r1.0'}.get(machine_type, 'unknown')
-        serial_number = get_serial_number_from_port(self.port)
-        pcb_rev = self.query(b'v', '<B')[0] if v_major > 22 else None
-
-        # log hardware information
-        logger.info(f'Bpod Finite State Machine {machine_str}')
-        logger.info(f'Serial number {serial_number}') if serial_number else None
-        logger.info(f'PCB revision {pcb_rev}') if pcb_rev else None
-        logger.info('Firmware version {}.{}'.format(*version))
-
-        # get hardware self-description
-        info: list[Any] = [
-            serial_number,
-            version,
-            machine_type,
-            machine_str,
-            pcb_rev,
-        ]
-        info.extend(self.query(b'H', '<2H6B'))
-        info.extend(self.read(f'<{info[-1]}s1B'))
-        info.extend(self.read(f'<{info[-1]}s'))
-        self.info = Bpod._Info(*info)
-
+    def _configure_channels(self) -> None:
         def collect_channels(description: bytes, dictionary: dict, channel_cls: type):
             """
             Generate a collection of Bpod channels.
@@ -338,70 +266,134 @@ class Bpod(SerialSingleton):
         logger.debug('Configuring I/O ports')
         input_dict = {b'B': 'BNC', b'V': 'Valve', b'P': 'Port', b'W': 'Wire'}
         output_dict = {b'B': 'BNC', b'V': 'Valve', b'P': 'PWM', b'W': 'Wire'}
-        collect_channels(self.info.input_description_array, input_dict, Input)
-        collect_channels(self.info.output_description_array, output_dict, Output)
+        collect_channels(self._hardware_config.input_description, input_dict, Input)
+        collect_channels(self._hardware_config.output_description, output_dict, Output)
 
-        # logger.debug("Configuring modules")
-        # self.modules = Modules(self)
+    def _detect_additional_serial_ports(self) -> None:
+        """Detect additional USB-serial ports."""
+        # First, assemble a list of candidate ports
+        candidate_ports = [
+            p.device for p in comports() if p.vid in VIDS_BPOD and p.device != self.port
+        ]
 
-    def close(self):
-        """Disconnect the state machine and close the serial connection."""
-        if not self.is_open:
-            return
-        logger.debug('Disconnecting state machine')
-        self.write(b'Z')
-        super().close()
+        # Exclude those devices from the list that are already sending a discovery byte
+        for port in candidate_ports:
+            try:
+                with ExtendedSerial(port, timeout=0.15) as ser:
+                    if ser.read(1) == bytes([222]):
+                        candidate_ports.remove(port)
+            except SerialException:
+                pass
 
-    def handshake(self, raise_exception_on_fail: bool = True) -> bool:
+        # Find second USB-serial port
+        for port in candidate_ports:
+            try:
+                with ExtendedSerial(port, timeout=0.15) as ser:
+                    self.serial0.write(b'{')
+                    if ser.read(1) == bytes([222]):
+                        ser.reset_input_buffer()
+                        ser.timeout = None
+                        self.serial1 = ser
+                        candidate_ports.remove(port)
+                        break
+            except SerialException:
+                pass
+
+        # State Machine 2+ uses a third USB-serial port
+        if self.version.machine == 4:
+            for port in candidate_ports:
+                try:
+                    with ExtendedSerial(port, timeout=0.15) as ser:
+                        self.serial0.write(b'}')
+                        if ser.read(1) == bytes([223]):
+                            ser.reset_input_buffer()
+                            ser.timeout = None
+                            self.serial2 = ser
+                            break
+                except SerialException:
+                    pass
+
+    def _handshake(self):
         """
-        Try to perform handshake with Bpod device.
+        Perform a handshake with the Bpod.
 
-        Returns
-        -------
-        bool
-            True if successful, False otherwise.
-
-        Notes
-        -----
-        This will reset the state machine's session clock and flush the serial port.
+        Raises
+        ------
+        BpodException
+            If the handshake fails.
         """
         try:
-            return self.query(b'6') == b'5'
+            self.serial0.timeout = 0.2
+            if not self.serial0.verify(b'6', b'5'):
+                raise BpodError(f'Handshake with device on {self.port} failed')
+            self.serial0.timeout = None
         except SerialException as e:
-            if raise_exception_on_fail:
-                raise BpodException('Handshake failed') from e
+            raise BpodError(f'Handshake with device on {self.port} failed') from e
         finally:
-            self.reset_input_buffer()
+            self.serial0.reset_input_buffer()
+        logger.debug(f'Handshake with Bpod on {self.port} successful')
 
-        if raise_exception_on_fail:
-            raise BpodException('Handshake failed')
-        return False
+    @property
+    def port(self) -> str | None:
+        """The port of the Bpod's primary serial device."""
+        return self.serial0.port
+
+    @property
+    def version(self) -> VersionInfo:
+        """Version information of the Bpod's firmware and hardware."""
+        return self._version
+
+    def open(self):
+        """
+        Open the connection to the Bpod.
+
+        Raises
+        ------
+        SerialException
+            If the port could not be opened.
+        BpodException
+            If the handshake fails.
+        """
+        if self.serial0.is_open:
+            return
+        self.serial0.open()
+        self._handshake()
+
+    def close(self):
+        """Close the connection to the Bpod."""
+        if hasattr(self, 'serial0') and self.serial0.is_open:
+            self.serial0.write(b'Z')
+            self.serial0.close()
 
     def update_modules(self):
-        pass
-        # self.write(b"M")
+        self.serial0.write(b'M')
         # modules = []
         # for i in range(len(modules)):
-        #     if self.read() == bytes([1]):
+        #     if self.serial0.read() == bytes([1]):
         #         continue
-        #     firmware_version = self.read(4, np.uint32)[0]
-        #     name = self.read(int(self.read())).decode("utf-8")
+        #     firmware_version = self.serial0.read(4, np.uint32)[0]
+        #     name = self.read(int(self.serial0.read())).decode('utf-8')
         #     port = i + 1
         #     m = Module()
-        #     while self.read() == b"\x01":
-        #         match self.read():
-        #             case b"#":
-        #                 number_of_events = self.read(1, np.uint8)[0]
-        #             case b"E":
-        #                 for event_index in range(self.read(1, np.uint8)[0]):
-        #                     l_event_name = self.read(1, np.uint8)[0]
-        #                     module["events"]["index"] = event_index
-        #                     module["events"]["name"] = self.read(l_event_name, str)[0]
+        #     while self.serial0.read() == b'\x01':
+        #         match self.serial0.read():
+        #             case b'#':
+        #                 number_of_events = self.serial0.read(1, np.uint8)[0]
+        #             case b'E':
+        #                 for event_index in range(self.serial0.read(1, np.uint8)[0]):
+        #                     l_event_name = self.serial0.read(1, np.uint8)[0]
+        #                     module['events']['index'] = event_index
+        #                     module['events']['name'] = self.serial0.read(
+        #                         l_event_name, str
+        #                     )[0]
         #         modules[i] = module
         #     self._children = modules
+        pass
 
 
 class Channel(ABC):
+    """Abstract base class representing a channel on the Bpod device."""
+
     @abstractmethod
     def __init__(self, bpod: Bpod, name: str, io_type: bytes, index: int):
         """
@@ -421,14 +413,17 @@ class Channel(ABC):
         self.name = name
         self.io_type = io_type
         self.index = index
-        self._query = bpod.query
-        self._write = bpod.write
+        self._query = bpod.serial0.query
+        self._write = bpod.serial0.write
+        self._validate_response = bpod.serial0.verify
 
     def __repr__(self):
         return self.__class__.__name__ + '()'
 
 
 class Input(Channel):
+    """Input channel class representing a digital input channel."""
+
     def __init__(self, *args, **kwargs):
         """
         Input channel class representing a digital input channel.
@@ -449,7 +444,8 @@ class Input(Channel):
         bool
             True if the input channel is active, False otherwise.
         """
-        return self._query(['I', self.index], 1) == b'\x01'
+        return self._validate_response(b'I' + bytes([self.index]), b'\x01')
+        # return self._query(['I', self.index], 1) == b'\x01'
 
     def override(self, state: bool) -> None:
         """
@@ -460,7 +456,7 @@ class Input(Channel):
         state : bool
             The state to set for the input channel.
         """
-        self._write(['V', state])
+        self._write([b'V', state])
 
     def enable(self, state: bool) -> None:
         """
@@ -475,6 +471,8 @@ class Input(Channel):
 
 
 class Output(Channel):
+    """Output channel class representing a digital output channel."""
+
     def __init__(self, *args, **kwargs):
         """
         Output channel class representing a digital output channel.
@@ -498,48 +496,4 @@ class Output(Channel):
         """
         if isinstance(state, int) and self.io_type in (b'D', b'B', b'W'):
             state = state > 0
-        self._write(['O', self.index, state.to_bytes(1, 'little')])
-
-
-class Module:
-    pass
-
-
-def find_bpod_ports() -> Iterator[str]:
-    """
-    Discover serial ports used by Bpod devices.
-
-    This method scans through the list of available serial ports and identifies ports
-    that are in use by a Bpod device. It does so by briefly opening each port and
-    checking for a specific byte pattern (byte 222). Ports matching this pattern are
-    yielded.
-
-    Yields
-    ------
-    str
-        The names of available serial ports compatible with the Bpod device.
-
-    Notes
-    -----
-    The method employs a brief timeout when opening each port to minimize the impact on
-    system resources.
-
-    SerialException is caught and ignored, allowing the method to continue scanning even
-    if certain ports encounter errors during opening.
-
-    Examples
-    --------
-    .. code-block:: python
-
-        for port in Bpod.find():
-            print(f'Bpod on {port}')
-        # Bpod on COM3
-        # Bpod on COM6
-    """
-    for port in (p for p in list_ports.comports() if p.vid == VID_TEENSY):
-        try:
-            with serial.Serial(port.device, timeout=0.2) as ser:
-                if ser.read(1) == bytes([222]):
-                    yield port.device
-        except serial.SerialException:
-            pass
+        self._write([b'O', self.index, bytes([state])])