+++ CHANGELOG.md
@@ -5,3 +5,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.1.0-alpha.1]
+
+### Changed
+
+- Major reorganization of the existing codebase.
+- Renamed `serial_extensions` module to `serial`.
+
+### Removed
+
+- Removed `SerialSingleton` class.
+
+## [0.1.0-alpha] - 2023-03-05
+
+First alpha release. Nothing works.
+
+[0.1.0-alpha.1]: https://github.com/int-brain-lab/bpod-core/releases/tag/0.1.0-alpha.1
+[0.1.0-alpha]: https://github.com/int-brain-lab/bpod-core/releases/tag/0.1.0-alpha